services/secret_manager/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
asyncpg==0.29.0
nats-py==2.6.0
python-dotenv==1.0.0
